code/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
google-cloud-bigquery==3.0.1
